forked and starting to make a roll down path game
forked from hackaball.  converted Cylinder to a prefab.  added second
cylinder to start the path down.  added CameraController to have the
camera follow a game object.  The CameraController is very simple at
this point, it simply follows an object based on an offset x,y,z.  In
PlayerController add more public variables to all specification of the
spawn point.